Rework all tests to work with current hackunit version
