+++ src/Definition/BackgroundCode.php
@@ -1,0 +1,25 @@
+<?hh // strict
+
+namespace kilahm\Clio\Definition;
+
+enum BackgroundCode : int as int
+{
+    normal     = 49;
+    black      = 40;
+    red        = 41;
+    green      = 42;
+    yellow     = 43;
+    blue       = 44;
+    magenta    = 45;
+    cyan       = 46;
+    light_gray = 47;
+
+    dark_gray     = 100;
+    light_red     = 101;
+    light_green   = 102;
+    light_yellow  = 103;
+    light_blue    = 104;
+    light_magenta = 105;
+    light_cyan    = 106;
+    white         = 107;
+}
