+++ src/Definition/CliTextAlign.php
@@ -1,0 +1,10 @@
+<?hh // strict
+
+namespace kilahm\Clio\Definition;
+
+enum CliTextAlign : string
+{
+    Left   = 'left';
+    Right  = 'right';
+    Center = 'center';
+}
