+++ src/Definition/ForegroundCode.php
@@ -1,0 +1,26 @@
+<?hh // strict
+
+namespace kilahm\Clio\Definition;
+
+enum ForegroundCode : int as int
+{
+    normal     = 39;
+    black      = 30;
+    red        = 31;
+    green      = 32;
+    yellow     = 33;
+    blue       = 34;
+    magenta    = 35;
+    cyan       = 36;
+    light_gray = 37;
+
+    dark_gray     = 90;
+    light_red     = 91;
+    light_green   = 92;
+    light_yellow  = 93;
+    light_blue    = 94;
+    light_magenta = 95;
+    light_cyan    = 96;
+    white         = 97;
+
+}
